+++ db/migrate/20191117204452_add_slug_to_characters.rb
@@ -1,0 +1,8 @@
+class AddSlugToCharacters < ActiveRecord::Migration[5.2]
+  def change
+    add_column :characters, :slug, :string
+    add_index :characters, :slug
+
+    Character.find_each(&:save)
+  end
+end